cli: init cmd migrated to docopts.
